Implement node list matching logic only once
This seems like a bad change at the moment, since we have to copy the
slices. However, we want to add complexity to the list matching logic
soon, like matching any number of nodes. We want to add the complexity
in one place, not four.

If the allocations are ever a bottleneck, we can re-use the underlying
arrays.

Updates #2.