rework restarting of list matches
The issue was with how we were keeping track of the matches made so far.
In the test cases added, we would try for $*_ to first match "", then
"a", and then "a, b", the last of which should work.

But since we were calling m.node to match $x against each of the
elements following those lists, the matcher would remember the first
forever - "a" - and thus refuse to accept "b" or "c".

Fix this by also being able to restart the map that keeps track of such
final matches. This makes sure that nested m.node calls will be undone
if we need them to.

This also means we can get rid of the old "recorded" logic, as it was
trying (quite badly) to accomplish the same.

Fixes #18.