allow "b; c" to match "{ a; b; c; d; }"
Do this only for statements. See the issue for the reasoning.

This also lets us get rid of the list glue in matches. While at it, also
deduplicate the node results. We can't use ast.Node as a key as some are
slices, and slices aren't comparable (and thus not valid map keys).

Fixes #20.