give variables (wildcards) numeric IDs from 0
Up until now, we kept extra info on them (such as the * for "any number
of") in the ident name itself, via more prefixes like wildExtraAny.

This doesn't scale well to more variations, though. And especially for
the upcoming regex matching, since storing regex expressions or pointers
in such an ident encoding would be hard.

Instead, give them an ID so that we can keep information on them
separately, including the original name of the variable.