add typed loader tests
As expected, stuff was broken as we were bypassing the loader in the
unit tests. We weren't actually collecting the types.Info in the loader.

To do so, refactor our list of nodes into a list of packages. Now the
matches are done per-package, instead of all at once.

While at it, capture type errors better so that they don't go to stderr
on their own.