add ability to filter the underlying type class
This way, it's possible to say "I want this to be an array" or "I want
this to be a pointer".

Still not possible to say "a pointer to an array", but this is a good
first step. Perhaps a way to do it would be stack-like, such as the
following for the first example:

	is(array pointer)