remove assignability TODO in match_test
go/types is right; untyped int can be converted to int, but not assigned
directly. As the spec says, an untyped constant can only be assigned to
a variable of type T if the constant is representable by a value of type
T. That is false in this case, as untyped integers don't always fit in
an int.