rework dollar expressions and some commands
Get rid of the $() notation. Instead, split these "wildcard operators"
into their own commands, which simplifies the syntax and the code.

With this change, the following query:

	-x '$(x /elems/ type([]string))'

Becomes:

	-x '$x' -a 'rx(elems)' -a 'type([]string)'

This will also allow these commands to be more composable and powerful.
For example, now one can check for type equality on all expressions, not
just dollar expressions.

While at it, also add a -p command to navigate up a number of parent
nodes, which is required now that queries tend to navigate to
sub-expressions more often.

Fixes #28.