don't crash on types that cannot be found
For now, simply ignore them. Erroring is not an option, as other
sub-matches might have found a type.

One option was to find the type globally across the entire program. But
that would be a slightly less precise approach, as this one finds the
type from the scope of the submatch. That will work better when the type
in question is defined in a local scope such as a func body.

I might revisit this mechanism in the future.

Fixes #30.