all: port to go/packages
To get module support, mainly. We also get rid of the deprecated
go/loader dependency.

For the sake of simplicity, the untyped loader has been removed for now.
We might re-add them in the future via LoadFiles.

This isn't perfect, so we'll probably clean up the code in the following
days. But at least 'go test' now passes on Go 1.11, and modules seem to
work.

Fixes #39.