don't load test packages by default
In go/packages, this currently means multiplying the loading and
typechecking work most of the time. Not only because of the _test.go
files and their extra dependencies, but also because we need to load and
typecheck internal test packages.

However, most queries are geared towards non-test code, so it's best not
to assume that the user wants to include test code. This way, the tool
is simpler and faster by default. It's also worth noting that most
static analysis tools already skip tests by default.

Reduces 'go test' time from ~2.4s to ~0.4s on my laptop.

We might want to reconsider the performance cost if
https://github.com/golang/go/issues/29258 is implemented, or if gogrep
starts being used with a language server.