support matching struct types and fields
The first piece of the puzzle was simple enough; adding support for
matching type specs.

The second piece was more subtle, which was to allow $var to match an
entire field. The tricky part is that our placeholder identifier gets
parsed as an anonymous struct field, as a type expression.

The fix was thus to catch that case specifically in ast.Field and in the
code that extracts wildcard information from a node.

Finally, add tests.

Fixes #38.