Handle `*ast.KeyValueExpr` in `fromWildNode`
Previously following match tests will get inconsistent result:

- `{[]string{"-x", "struct{a int}{a: $_}"}, "struct{a int}{a: 1}", 1}`: OK
- `{[]string{"-x", "struct{a int}{a: $*_}"}, "struct{a int}{a: 1}", 1}`: NOK

This is a quick fix while not sure if this makes sense. The cause why
original code doesn't work for the 2nd test case is because when
comparing `expr` (`*ast.Ident{Name: "gogrep_0"}`) and `node`
(`*ast.BasicLit{Value: "1"}`) in `matcher.node()`, it will return
`false` simply because the wildcard is `any`. I'm not sure the design
behind this, so...